backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
decorator==5.1.1
defusedxml==0.8.0rc2
Django==3.2
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
executing==2.0.1
idna==3.7
iniconfig==2.0.0
ipython==8.25.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
oauthlib==3.2.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pip-check==2.8.1
pluggy==0.13.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
short_url==1.2.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.0
stack-data==0.6.3
terminaltables==3.1.10
toml==0.10.2
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.11.1
